integration with middleware, supported context in scoreboard
