+++ pages/ScoreBoard/SubBlock/Leaderboard.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { Grid, Typography, Stack } from "@mui/material";
 import { ClanCard } from "../../../components/RankCard";
+import { AppContext } from "../AppContext";
 
 //for data config related matters:
 import properties from "../../../config/prop-config.json";
@@ -9,41 +10,9 @@ import properties from "../../../config/prop-config.json";
 import { Reorder } from "framer-motion";
 
 export default function LeaderBoard() {
-  const [loading, setLoading] = React.useState(true);
-  const [data, setData] = React.useState([]);
-  const [result, setResult] = React.useState([]);
+  const { data, loading, result, setResult } = React.useContext(AppContext);
 
-  const fetchData = async () => {
-    try {
-      const response = await fetch(
-        "https://script.google.com/macros/s/AKfycbzsp75jnaDAQ5EvYUJ5FUejFZpNrtEGc_tdkErdbwtIgUHCHjM48-iKLzLAke2b16E1Cg/exec"
-      );
-      const data = await response.json();
-      const res = Object.entries(data).map(([name, obj]) => ({ name, ...obj }));
-      setData(data); // Set the fetched data into the state
-      const sortedData = [...res];
-      sortedData.sort((a, b) => b.clan - a.clan);
-      setResult(sortedData.map((sortedData) => sortedData.name));
-      setLoading(false); // Set loading to false once data is fetched
-    } catch (error) {
-      console.error("Error fetching data:", error);
-      setLoading(false); // Set loading to false even if there's an error
-    }
-  };
-
-  React.useEffect(() => {
-    fetchData();
-
-    // Fetch data at regular intervals using setInterval
-    const intervalId = setInterval(fetchData, 2000);
-
-    // Clean up the interval when the component unmounts to prevent memory leaks
-    return () => {
-      clearInterval(intervalId);
-    };
-  }, []);
-
-  console.log(result);
+  console.log(data);
 
   return (
     <>
